Add test vectors extender.
