+++ changelogs/fragments/16-create-vectors.yml
@@ -1,0 +1,2 @@
+minor_changes:
+  - "Create script to update/extend the test vectors automaticaly (https://github.com/ansible-community/antsibull-docs-parser/pull/16)."
